+++ Fields/Interfaces/IField.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Numerics;
 using System.Text;
 using System.Xml.Linq;
 
@@ -7,13 +8,13 @@ namespace BLS.Fields.Interfaces
 {
     public interface IField<TElement> where TElement : IFieldElement<TElement>
     {
-        int Characteristic { get; }  // p
+        BigInteger Characteristic { get; }  // p
         int ExtensionDegree { get; } // 1 for Fp, k for Fp^k
 
         // Canonical elements
         TElement Zero { get; }
         TElement One { get; }
-        TElement FromInt(long value);
+        TElement FromInt(BigInteger value);
         bool IsValid(TElement x);
     }
 }
